+++ NEWS.md
@@ -15,6 +15,7 @@
 * Changed function names: `ShowPanglaoDBMeta` to `ExtractPanglaoDBMeta`, `PrepareZenodo` to `ExtractZenodoMeta`.
 * Supported downloadding normal (non-10x) bam file in `DownloadBam`.
 * `ExtractZenodoMeta` supported a vector of Zenodo dois.
+* Added `fastq.type` in `SplitSRA` to deal with fastq files from 10x, other scRNA-seq protocols and bulk RNA-seq.
 
 -------------------